Refactor: Extract manufacturing and picking configuration logic into dedicated command classes for better modularity and maintainability.
