+++ backend/de.metas.frontend-testing/src/main/java/de/metas/frontend_testing/masterdata/mobile_configuration/JsonMobileConfigResponse.java
@@ -1,6 +1,7 @@
 package de.metas.frontend_testing.masterdata.mobile_configuration;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import de.metas.handlingunits.picking.config.mobileui.PickAttribute;
 import de.metas.handlingunits.picking.config.mobileui.PickToStructure;
 import de.metas.handlingunits.picking.config.mobileui.PickingJobAggregationType;
 import de.metas.handlingunits.picking.job.service.CreateShipmentPolicy;
@@ -38,6 +39,7 @@ public class JsonMobileConfigResponse
 		@Nullable Boolean alwaysSplitHUsEnabled;
 		@Nullable Boolean shipOnCloseLU;
 		@Nullable Set<PickToStructure> pickTo;
+		@Nullable Set<PickAttribute> readAttributes;
 		@Nullable Boolean filterByQRCode;
 		@Nullable Boolean allowCompletingPartialPickingJob;
 		@Nullable Boolean isAnonymousPickHUsOnTheFly;
